Create index.html
An index.html is a universally understood gateway to a website. It will be understood by an 'internet browser' (firefox, chrome, edge, safari) to be the place where a website begins. This has worked since the early days of the web and will do into the future of the internet!